pyproject.toml
[project]
name = "bali-flood-prevention"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.0",
    "datasets>=2.19",
    "fastapi>=0.115",
    "gguf>=0.10",
    "huggingface_hub>=0.23",
    "httpx>=0.27",
    "modal>=1.4.2",
    "num2words>=0.5",
    "openai>=1.0",
    "pandas>=2.2",
    "peft>=0.14",
    "pillow>=10.0",
    "protobuf>=5",
    "python-dotenv>=1.0",
    "requests>=2.32",
    "sentencepiece>=0.2",
    "streamlit>=1.37",
    "torch>=2.3",
    "torchvision>=0.26.0",
    "transformers>=4.47,<5",
    "tqdm>=4.66",
    "uvicorn>=0.30",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bali_flood_prevention"]

[tool.pyright]
typeCheckingMode = "strict"
